% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_String.R
\name{Clean_String}
\alias{Clean_String}
\title{Una funcion que limpia una cadena de caracteres, removiendo puntuacion y numeros y tokenizandola}
\usage{
Clean_String(str)
}
\arguments{
\item{str}{Una cadena de caracteres de entrada tal como "This is a cool function!"}
}
\value{
Un vector que contiene todos los tokens validos de la cadena de caracteres de entrada original
}
\description{
Una funcion que limpia una cadena de caracteres, removiendo puntuacion y numeros y tokenizandola
}
